Peer review

+++ assignment4/blur_1.py
@@ -1,3 +1,11 @@
+'''A module that blurs images using a 3x3 averaging kernel.
+
+Can be run from the command line to blur a specified image.
+Uses simple python-loops as its implementation.
+'''
+
+import sys
+import os
 import numpy as np
 import cv2
 
@@ -16,21 +24,29 @@ def blur_image_python(image):
     width -= 2
     bimage = np.empty([height, width, channels], dtype=np.uint8)
 
-    dh = [1, 1, 1, 0, 0, 0, -1, -1, -1]
-    dw = [1, 0, -1, 1, 0, -1, 1, 0, -1]
+    height_offsets = [1, 1, 1, 0, 0, 0, -1, -1, -1]
+    width_offsets = [1, 0, -1, 1, 0, -1, 1, 0, -1]
 
     for h in range(height):
         for w in range(width):
             for c in range(channels):
                 intensity = 0.
                 for i in range(9):
-                    intensity += image_[h + dh[i] + 1, w + dw[i] + 1, c]
+                    intensity += image_[h + height_offsets[i] + 1, \
+                    w + width_offsets[i] + 1, c]
                 bimage[h, w, c] = intensity/9.
 
     return bimage
 
 
 if __name__ == '__main__':
-    image = cv2.imread('beatles.jpg')
-    image_ = blur_image_python(image)
-    cv2.imwrite('blurred_image.jpg', image_)
+    if len(sys.argv) == 2:
+        inputfile = sys.argv[1]
+        if os.path.isfile(inputfile):
+            image = cv2.imread(inputfile)
+            image_ = blur_image_python(image)
+            cv2.imwrite('blurred_image.jpg', image_)
+        else:
+            print(f'File {inputfile} does not exist.')
+    else:
+        print('Usage: blur_1.py inputfile')
